dimensions

+++ _scripts/problems/classes/actions.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from problems.classes.problem import Problem
 from enum import Enum
+from decimal import Decimal
 
 class ActionType(Enum):
     STRETCH = 0
@@ -11,7 +12,7 @@ class Action:
     problem: Problem
     action_type: ActionType
     node: str
-    distance: float
+    distance: Decimal
     succesful: bool = False
 
     def __repr__(self) -> str:
